+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 All notable changes will be tracked in this change log.  This project uses [Semantic Versioning](http://semver.org/)
 
+## Unreleased
+### Fix
+
+- When clearing log file, if that fails attempt to delete the file.
+
 ## [0.11.1] - 2015-06-22
-## Fix
+### Fix
 
 - Added support for when toLocaleDateString and toLocaleTimeString are not supported.  Will instead use toDateString and toTimeString then format date/time accordingly.
 - Added support for 401 errors what debug log update checker is running.  A 401 error will not stop the auto checker.